Oil Rig
Oil Rigs now works and you get oil from it over time. Next i'll make a ui for all the buildings that'll display it's resources over time amount and a button to destroy it, and after that i'll make the resources that you farm will get destroyed if you harvest it for to long.