AI improvments
Made the AI better while moving to a position, and added the soldier modle i bought to the basic troops. I'm still working on A* pathfinding, and next i'm going to finish implementing more of the assets i got and make a way for soldiers to do more damage to air units, air units to do more damage to tank units and tank units to do more damage to soldier units.